Split steps into individual modules
