+++ packages/core/src/steps/index.ts
@@ -1,5 +1,13 @@
-export * from "./command";
-export * from "./git";
-export * from "./package";
-export * from "./publish";
-export * from "./version";
+export * from "./command/run";
+export * from "./git/commit";
+export * from "./git/commitChanges";
+export * from "./git/configureGit";
+export * from "./git/createReleaseBranch";
+export * from "./git/push";
+export * from "./git/stageFiles";
+export * from "./package/resolvePackage";
+export * from "./publish/publish";
+export * from "./version/bumpVersion";
+export * from "./version/calculateNextVersion";
+export * from "./version/resolvePublishedVersion";
+export * from "./version/writePackageVersion";
