+++ packages/core/src/steps/version/types.ts
@@ -1,0 +1,7 @@
+import type { ReleaseLevel, VersionResolver } from "../../types/RlseConfig";
+
+export type VersionOptions = {
+  level?: ReleaseLevel;
+  pre?: boolean;
+  version?: string | VersionResolver;
+};
